indexer/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
PyPDF2
tqdm
numpy
python-multipart